pyproject.toml
[project]
name = "erpnext-mcp"
version = "0.1.0"
description = "MCP Server for ERPNext REST API"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
authors = [
    {name = "Ching-Tech"},
]
keywords = ["mcp", "erpnext", "frappe", "erp", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "fastmcp>=2.0.0",
    "httpx>=0.27.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=0.24",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.urls]
Homepage = "https://github.com/ching-tech/erpnext-mcp"
Repository = "https://github.com/ching-tech/erpnext-mcp"
Issues = "https://github.com/ching-tech/erpnext-mcp/issues"

[project.scripts]
erpnext-mcp = "erpnext_mcp.server:main"
